LayoutTests/platform/gtk/accessibility/custom-elements/autocomplete-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("custom-list").role is "AXRole: AXComboBox"
FAIL accessibilityController.accessibleElementById("custom-list").stringAttributeValue("AXAutocompleteValue") should be list. Was .
PASS accessibilityController.accessibleElementById("custom-textbox").role is "AXRole: AXTextField"
FAIL accessibilityController.accessibleElementById("custom-textbox").stringAttributeValue("AXAutocompleteValue") should be inline. Was .
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
